STRWEB-76: Replace babel loader with esbuild loader
