+++ package.json
@@ -38,9 +38,9 @@
     "core-js": "^3.6.1",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.4.0",
-    "css-minimizer-webpack-plugin": "^4.2.2",
     "csv-loader": "^3.0.3",
     "debug": "^4.0.1",
+    "esbuild-loader": "^3.0.1",
     "express": "^4.14.0",
     "favicons": "^7.0.1",
     "favicons-webpack-plugin": "^6.0.0-alpha.1",
@@ -73,7 +73,6 @@
     "svgo": "^1.2.2",
     "svgo-loader": "^2.2.1",
     "tapable": "^1.0.0",
-    "terser-webpack-plugin": "^5.3.5",
     "ts-loader": "^9.4.1",
     "typescript": "^4.2.4",
     "url-loader": "^4.1.1",
