+++ test/webpack/esbuild-loader-rule.spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const babelLoaderRule = require('../../webpack/babel-loader-rule');
+const babelLoaderRule = require('../../webpack/esbuild-loader-rule');
 
 describe('The babel-loader-rule', function () {
   describe('test condition function', function () {
